1) API : to show SOURCE_TYPE on the basis of usertype. 2)'Source' And 'Request type' interp in stolen recovery . 3) implemented spring cache for config data. 4) API : Add stateInterp in Stock/Stolen. 5) Show Stock/Stolen on the basis of states configured on state_mgmt and updated version 1.0.10. 6) TableActions API ready. 7) Updated created on date format : yyyy-MM-dd HH:mm 8) Added usertype check in consignment delete API. 9 Added not empty check for txnId filter on stock/Stolen.

+++ config/src/main/java/com/gl/ceir/config/repository/UserProfileRepository.java
@@ -7,6 +7,6 @@ import com.gl.ceir.config.model.UserProfile;
 public interface UserProfileRepository extends JpaRepository<UserProfile, Integer> {
 
 
-	public UserProfile getByUserId(int userid);
+	public UserProfile getByUserId(long userid);
 
 }
